requirements.txt
flask==2.3.2
Flask-SQLAlchemy==3.0.2
sqlalchemy==2.0.16
psycopg2-binary==2.9.6
python-dotenv==1.0.0
flask_wtf==1.1.1
WTForms==3.0.1
Flask-Migrate==4.0.4
flask-bootstrap==3.3.7.1